Change fixed-point integer format from 2.6 to 1.7
Previously, we used 2.6 -- signed 8 bit format with range (-2,2). 
However, we only need to represent values in the range [-1,+1]. Using 
1.7 gives us 1 extra bit of, badly needed, resolution. The downside is 
that we can represent -1, but not +1. The largest value is 127/128 = 
0.9922.

This is a shame, as 1.0 comes up quite frequently, especially in unit 
vectors. However, the extra bit of resolution makes the difference 
between the ships turning fairly smoothly with stable panels rather than 
jerkily with panels that seem to wobble, due to the lack of accuracy of 
the fractional values. For this, it is worth the trade off of making 
anything that is multiplied by a unit vector < 1% smaller.